@@ -1,6 +1,6 @@
 {
     "name": "opencapture",
-    "version": "2.5.0",
+    "version": "2.5.1",
     "scripts": {
         "lint": "ng lint",
         "check-update": "ncu",
@@ -13,58 +13,58 @@
     "private": true,
     "dependencies": {
         "@angular-material-extensions/select-country": "^7.0.1",
-        "@angular/animations": "~14.2.4",
-        "@angular/cdk": "^14.2.3",
-        "@angular/common": "~14.2.4",
-        "@angular/compiler": "~14.2.4",
-        "@angular/core": "~14.2.4",
-        "@angular/forms": "~14.2.4",
-        "@angular/material": "^14.2.3",
-        "@angular/material-moment-adapter": "^14.2.3",
-        "@angular/platform-browser": "~14.2.4",
-        "@angular/platform-browser-dynamic": "~14.2.4",
-        "@angular/router": "~14.2.4",
+        "@angular/animations": "~14.2.8",
+        "@angular/cdk": "^14.2.6",
+        "@angular/common": "~14.2.8",
+        "@angular/compiler": "~14.2.8",
+        "@angular/core": "~14.2.8",
+        "@angular/forms": "~14.2.8",
+        "@angular/material": "^14.2.6",
+        "@angular/material-moment-adapter": "^14.2.6",
+        "@angular/platform-browser": "~14.2.8",
+        "@angular/platform-browser-dynamic": "~14.2.8",
+        "@angular/router": "~14.2.8",
         "@biesbjerg/ngx-translate-extract-marker": "^1.0.0",
         "@fortawesome/fontawesome-free": "^6.2.0",
         "@ngx-translate/core": "^14.0.0",
         "@ngx-translate/http-loader": "^7.0.0",
-        "@schematics/angular": "^14.2.4",
+        "@schematics/angular": "^14.2.7",
         "@swimlane/ngx-charts": "^20.1.0",
-        "@syncfusion/ej2": "^20.3.47",
+        "@syncfusion/ej2": "^20.3.52",
         "jquery": "^3.6.1",
         "moment": "^2.29.4",
         "ng-click-outside2": "^11.1.0",
         "ng-sortgrid": "^14.0.0",
         "ng2-datepicker": "^12.0.0",
         "ngx-file-drag-drop": "^6.0.0",
-        "ngx-mat-select-search": "^4.2.1",
+        "ngx-mat-select-search": "^5.0.0",
         "remove-accents": "^0.4.3",
         "rxjs": "~7.5.7",
         "svg-country-flags": "^1.2.10"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^14.2.4",
-        "@angular-devkit/build-webpack": "^0.1402.4",
+        "@angular-devkit/build-angular": "^14.2.7",
+        "@angular-devkit/build-webpack": "^0.1402.7",
         "@angular-eslint/builder": "14.1.2",
         "@angular-eslint/eslint-plugin": "14.1.2",
         "@angular-eslint/eslint-plugin-template": "14.1.2",
         "@angular-eslint/schematics": "14.1.2",
         "@angular-eslint/template-parser": "14.1.2",
-        "@angular/cli": "~14.2.4",
-        "@angular/compiler-cli": "~14.2.4",
+        "@angular/cli": "~14.2.7",
+        "@angular/compiler-cli": "~14.2.8",
         "@bartholomej/ngx-translate-extract": "^8.0.2",
         "@types/d3-scale": "^4.0.2",
         "@types/d3-selection": "^3.0.3",
-        "@types/node": "^18.8.2",
-        "@typescript-eslint/eslint-plugin": "5.39.0",
-        "@typescript-eslint/parser": "5.39.0",
-        "autoprefixer": "^10.4.12",
-        "eslint": "^8.24.0",
+        "@types/node": "^18.11.8",
+        "@typescript-eslint/eslint-plugin": "5.41.0",
+        "@typescript-eslint/parser": "5.41.0",
+        "autoprefixer": "^10.4.13",
+        "eslint": "^8.26.0",
         "eslint-plugin-import": "latest",
         "eslint-plugin-jsdoc": "latest",
         "eslint-plugin-prefer-arrow": "latest",
-        "postcss": "^8.4.17",
-        "tailwindcss": "^3.1.8",
+        "postcss": "^8.4.18",
+        "tailwindcss": "^3.2.1",
         "ts-node": "^10.9.1",
         "typescript": "~4.8.4",
         "zone.js": "~0.11.8"
