@@ -17,18 +17,19 @@
 
 import { Injectable } from '@angular/core';
 import { AuthService } from "./auth.service";
-import { ActivatedRouteSnapshot, CanActivate, RouterStateSnapshot } from "@angular/router";
+import {ActivatedRoute, ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot} from "@angular/router";
 import { NotificationService } from "./notifications/notifications.service";
 import { TranslateService } from "@ngx-translate/core";
 import { UserService } from "./user.service";
 import { environment } from "../app/env";
-import { catchError, tap } from "rxjs/operators";
+import {catchError, finalize, tap} from "rxjs/operators";
 import { of } from "rxjs";
 import { HttpClient } from "@angular/common/http";
 import { LocaleService } from "./locale.service";
 import { ConfigService } from "./config.service";
 import { HistoryService } from "./history.service";
 import {LastUrlService} from "./last-url.service";
+import {LocalStorageService} from "./local-storage.service";
 
 @Injectable({
     providedIn: 'root'
@@ -36,6 +37,7 @@ import {LastUrlService} from "./last-url.service";
 export class LoginRequiredService implements CanActivate {
 
     constructor(
+        public router: Router,
         private http:HttpClient,
         private authService: AuthService,
         private userService: UserService,
@@ -45,22 +47,11 @@ export class LoginRequiredService implements CanActivate {
         private localeService: LocaleService,
         private historyService: HistoryService,
         private routerExtService: LastUrlService,
+        private localStorage: LocalStorageService
     ) {}
 
-    canActivate(): boolean {
-        const token = this.authService.getToken();
-        if (!token) {
-            this.translate.get('AUTH.not_connected').subscribe((translated: string) => {
-                const currentUrl = this.routerExtService.getCurrentUrl();
-                if (currentUrl !== '/logout' && currentUrl !== '/login' && currentUrl !== '/500') {
-                    this.authService.setCachedUrl(currentUrl.replace(/^\//g, ''));
-                }
-                this.notify.error(translated);
-                this.authService.logout();
-            });
-            return false;
-        }
-        if (this.userService.getUser() === undefined && this.userService.getUserFromLocal() === undefined) {
+    login(token: string, route: any) {
+        if ((this.userService.getUser() === undefined || Object.keys(this.userService.getUser()).length === 0) && this.userService.getUserFromLocal() === undefined) {
             this.http.post(
                 environment['url'] + '/ws/auth/login',
                 {
@@ -80,6 +71,10 @@ export class LoginRequiredService implements CanActivate {
                     this.notify.success(this.translate.instant('AUTH.authenticated'));
                     this.configService.readConfig().then(() => {
                         this.historyService.addHistory('general', 'login', this.translate.instant('HISTORY-DESC.login_with_token'));
+                        if (route) {
+                            this.localStorage.save('task_watcher_minimize_display', 'true');
+                            this.router.navigate([route]).then();
+                        }
                     });
                 }),
                 catchError((err: any) => {
@@ -92,6 +87,35 @@ export class LoginRequiredService implements CanActivate {
                 })
             ).subscribe();
         }
+    }
+
+    canActivate(): boolean {
+        const token = this.authService.getToken();
+        const tokenRouteStatic = this.authService.getTokenCustom('tokenRouteStatic');
+        let route = '';
+        if (!token) {
+            const params = new URLSearchParams(window.location.href);
+            let token_param = '';
+            for (const [key, value] of params.entries()) {
+                if (key === 'token') {
+                    token_param = value;
+                    route = window.location.hash.replace('#', '').replace(/\?.*/,'')
+                    this.login(token_param, route);
+                }
+            }
+        }
+        if (!token) {
+            this.translate.get('AUTH.not_connected').subscribe((translated: string) => {
+                const currentUrl = this.routerExtService.getCurrentUrl();
+                if (currentUrl !== '/logout' && currentUrl !== '/login' && currentUrl !== '/500') {
+                    this.authService.setCachedUrl(currentUrl.replace(/^\//g, ''));
+                }
+                this.notify.error(translated);
+                this.authService.logout();
+            });
+            return false;
+        }
+        this.login(token, tokenRouteStatic);
         return true;
     }
 }
